Fix LPeg precedence for Lua dot operators
LPeg doesn't have a "longest match is preferred" rule which means that
when the LPeg pattern "operator" in the Lua lexer first tries to match
the Lua dot operator and then tries to match the double dot operator,
the second match can never succeed because the first sub-pattern (the
single dot) has already been matched and will be returned as a
successful match. Because the single dot is already included in the last
sub-pattern (the set of single-character operators) I've just removed
the P '.' sub-pattern. I also removed the P prefixes because they're not
needed (as long as an LPeg pattern is on the left side of the long
expression of valid multi-character operators).